Learnt about Interfaces
Use interfaces as a way to categorize things *before* inheritance. E.g. interface called IDamageable, IDeathable, and IExplodable could be created before creating a class called Enemy that's to be inherited.

Another e.g. would be do use IDamageable and place that interface into the PlayerHealth class so the same methods for taking damage are used. This can be used when different gameobjects need to take damage to actively search for every object with the interface method instead of calling every object's method individually.